requirements.txt
fastmcp[v2.9.2]
kubernetes
fastapi
uvicorn[standard]
openai
asyncio
httpx
PyYAML